Added HDLRuby module of neural network and separated classes of FastNeurons.
・FastNeuronsのモデルごとにファイルを分割した。
・HDLRubyで書かれたニューラルネットワークのモジュールを追加した。
